# portals/common/core/static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "apk-site Admin / Release / Ops API",
    "version": "2.0.0",
    "description": "Auto-generated from Flask url_map (admin portal mode)."
  },
  "paths": {
...
    "/admin/approval": {
      "get": {
        "summary": "admin routes.admin approval page",
        "operationId": "admin_routes.admin_approval_page_get",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/api/ops-platform/actions/{trace_id}": {
      "get": {
        "summary": "project ops.ops platform action detail",
        "operationId": "project_ops.ops_platform_action_detail_get",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "name": "trace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/ops-platform/actions/approval": {
      "post": {
        "summary": "project ops.ops platform actions approval",
        "operationId": "project_ops.ops_platform_actions_approval_post",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        },
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/api/ops-platform/actions/execute": {
      "post": {
        "summary": "project ops.ops platform actions execute",
        "operationId": "project_ops.ops_platform_actions_execute_post",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        },
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/api/ops-platform/actions/validate": {
      "post": {
        "summary": "project ops.ops platform actions validate",
        "operationId": "project_ops.ops_platform_actions_validate_post",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        },
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
...
    "/api/ops-platform/flow-smoke": {
      "post": {
        "summary": "project ops.ops platform flow smoke",
        "operationId": "project_ops.ops_platform_flow_smoke_post",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        },
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
...
    "/api/ops-platform/stress-test": {
      "post": {
        "summary": "project ops.ops platform stress test",
        "operationId": "project_ops.ops_platform_stress_test_post",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        },
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
...
  }
}